src/requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
black==21.8b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cryptography==36.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
jwt==1.3.1
MarkupSafe==2.0.1
mercadopago==2.0.7
mypy-extensions==0.4.3
orjson==3.6.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.3.0
promise==2.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyMySQL==1.0.2
pyodbc==4.0.32
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
rsa==4.7.2
Rx==1.6.1
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
tomli==1.2.1
typing-extensions==3.10.0.2
ujson==4.1.0
urllib3==1.26.6
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
